+++ crates/streamling-e2e/tests/scan_sharing.rs
@@ -123,7 +123,8 @@ sinks:
         let missing: Vec<(i64,)> = ctx
             .postgres
             .query(&format!(
-                "SELECT s.id FROM generate_series(1, 100) AS s(id) \
+                // `generate_series(int, int)` yields INT4; the ids are BIGINT.
+                "SELECT s.id::bigint FROM generate_series(1, 100) AS s(id) \
                  LEFT JOIN public.{table} o ON o.id = s.id \
                  WHERE o.id IS NULL ORDER BY s.id"
             ))
